New: bump the internal block size.
This has two effects: first we compare more data in the quick compare stage, which will slow it down.

But when we get to the full compare stage we compare data much faster due to less overhead, especially when the underlying media is fast (SSD, cached in memory etc) since we have many fewer syscalls and channel writes etc.

In a test with 1.5G, total run time dropped from 2.1 to 1.7s.

Since the best value for this is filesystem dependent, the optimal value will vary depending on the system. But a value which is a multiple of both 512 and 4096 should be good for regular magnetic drives, while a higher value is good for SSD or RAM disks. 16 KB feels pretty reasonable.